% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{addValidationSuffix}
\alias{addValidationSuffix}
\title{Appends a validation suffix to a string.}
\usage{
addValidationSuffix(tagId)
}
\arguments{
\item{tagId}{ID of the tag.}
}
\value{
A string `"{tagId}-shinyreforms-validation"`.
}
\description{
Appends a validation suffix to a string.
}
