% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_input.R
\name{createHelpIcon}
\alias{createHelpIcon}
\title{Creates a shinyreforms help icon and pop-up.}
\usage{
createHelpIcon(helpText)
}
\arguments{
\item{helpText}{A tooltip to be displayed.}
}
\value{
A shiny div with an icon and pop-up tooltip.
}
\description{
Creates a shinyreforms help icon and pop-up.
}
