% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{getInputId}
\alias{getInputId}
\title{Returns an ID of an input tag.}
\usage{
getInputId(inputTag)
}
\arguments{
\item{inputTag}{A shiny tag to retrieve the ID from.}
}
\value{
ID of an input tag.
}
\description{
Returns an ID of an input tag.
}
