% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyreforms_css.R
\name{shinyReformsPage}
\alias{shinyReformsPage}
\title{Adds a shinyreforms dependency to a tag object.}
\usage{
shinyReformsPage(htmlTag)
}
\description{
Adds a shinyreforms dependency to a tag object.
}
