% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validated_input.R
\name{validatedInput}
\alias{validatedInput}
\title{Use to create shiny input tags with validation.
This should only be used in ShinyForm constructor.}
\usage{
validatedInput(tag, helpText = NULL, validators = c())
}
\description{
Use to create shiny input tags with validation.
This should only be used in ShinyForm constructor.
}
\examples{
shinyreforms::validatedInput(
    shiny::textInput("text_input", label="Username"),
    helpText="Username must have length between 4 and 12 characters.",
    validators=c(
        shinyreforms::ValidatorMinLength(4),
        shinyreforms::ValidatorMaxLength(12)
    )
)
}
