% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{Validator}
\alias{Validator}
\title{Class representing a Validator.}
\usage{
Validator(function(value) {...}, "Validation failed!")
}
\description{
Validators are used to validate input fields in a ShinyForm.
Validators are to be used with the validatedInput function.
A single input field can have several validators.
}
\details{
Package shinyformss defines a set of commonly used pre-defined
Validators. These include:
\describe{
 \item{ValidatorMinLength(minLength):}{Will fail if string is shorter than minLength.}
 \item{ValidatorMaxLength(maxLength):}{Will fail if string is longer than maxLength.}
 \item{ValidatorNonEmpty():}{Will fail if string is empty.}
}
}
\section{Fields}{

\describe{
\item{\code{test}}{Function returning a boolean value which will be used
to validate input.}

\item{\code{failMessage}}{Error message to display when validation fails.}
}}

