% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{ValidatorMaxLength}
\alias{ValidatorMaxLength}
\title{Validator requiring maximum length.}
\usage{
ValidatorMaxLength(maxLength)
}
\description{
Will return TRUE for strings longer than the maximum value.
}
