% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{ShinyForm}
\alias{ShinyForm}
\title{Class representing a ShinyForm form.}
\description{
Class representing a ShinyForm form.
}
\section{Fields}{

\describe{
\item{\code{id}}{Unique form id which can be used with Shiny input.}

\item{\code{elements}}{A list of ShinyForm input elements.}

\item{\code{validators}}{A list of Validators used to validate each field.}

\item{\code{submit}}{A submit Action button/link.}
}}

\section{Methods}{

\describe{
\item{ui}{Returns a list of shiny tags corresponding to the form.}
\item{validate}{Returns TRUE or FALSE based on from validation.}
\item{submissionEvent}{A form submission event to be used with shiny::observeEvent.}
\item{getValue}{Returns the value of a form element with a given id.}
}
}

