% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyform.R
\name{validatedInput}
\alias{validatedInput}
\title{Use to create shiny input tags with validation.
This should only be used in ShinyForm constructor.}
\usage{
validatedInput(tag, ...)
}
\description{
Use to create shiny input tags with validation.
This should only be used in ShinyForm constructor.
}
\examples{
shinyreforms::validatedInput(
    shiny::textInput("text_input", label="Username"),
    shinyreforms::ValidatorMinLength(4),
    shinyreforms::ValidatorMaxLength(12)
)
}
