% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{ValidatorMinLength}
\alias{ValidatorMinLength}
\title{Validator requiring minimum length.}
\usage{
ValidatorMinLength(minLength)
}
\arguments{
\item{minLength}{Minimum length of the input.}
}
\value{
A \code{Validator} checking that the input value is of length at
 least \code{minLength}.
}
\description{
Will return TRUE for strings longer than the minimum value.
}
