% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{ValidatorNonEmpty}
\alias{ValidatorNonEmpty}
\title{Validator requiring non-emptiness.}
\usage{
ValidatorNonEmpty()
}
\value{
A `Validator` to check if an input is non-empty.
}
\description{
The validator will return FALSE if the input is NULL, an empty
vector, or an empty string ("") and FALSE otherwise.
}
