% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{ValidatorRequired}
\alias{ValidatorRequired}
\title{Validator requiring a input (e.g. checkbox).}
\usage{
ValidatorRequired()
}
\description{
Will return FALSE if the input value is FALSE (e.g. like
for an unchecked textbox.)
}
